requirements.federation.txt
# === OMNI GENERATED REQUIREMENTS ===
# Generated: 2025-12-18 15:26:37.094689
# Uniqueness enforced by Omni Core

# Config parsing
    "rich>=13.0.0
# connectors and functions
    "openai>=1.99.0
# Data validation
# FastAPI framework
  "google-api-python-client>=2.157.0
# File system watching
    "pyyaml>=6.0
# For APIHubToolset.
  "aiosqlite>=0.21.0
# For CLI tools
  "fastapi>=0.115.0
# For MCP Session Manager
  "authlib>=1.5.1
# For RestAPI Tool
  "click>=8.1.8
# For SQLite database
  "anyio>=4.9.0
# go/keep-sorted start
  "PyYAML>=6.0.2
# Google API client discovery
  "google-cloud-aiplatform[agent_engines
# Pretty terminal output
    "pydantic>=2.0.0
# telemetry
    "opentelemetry-api>=1.24
# utilities
    "typing-extensions
<0.119.0
a2a-sdk>=0.2.8
ag-ui-protocol>=0.1.9
agent-framework
agent-framework-a2a
agent-framework-ag-ui
agent-framework-anthropic
agent-framework-azure-ai
agent-framework-azurefunctions
agent-framework-chatkit
agent-framework-copilotstudio
agent-framework-core
agent-framework-devui
agent-framework-lab
agent-framework-mem0
agent-framework-purview
agent-framework-redis
agno>=1.7.0
aiofiles==23.2.0
aiohttp==3.9.1
alembic==1.17.2
anthropic>=0.70.0
anyio==4.11.0
asyncio-extensions>=0.1.0
asyncio-mqtt==0.16.1
asyncpg==0.30.0
azure-ai-agents == 1.2.0b5
azure-ai-agentserver-agentframework==1.0.0b3
azure-ai-projects >= 2.0.0b1
azure-core>=1.30.0
azure-functions
azure-functions-durable
azure-identity>=1
azure-quantum>=1.0.0
bcrypt==4.1.2
black # For code formatting
black==23.12.1
browserbase==1.3.0
cirq>=1.2.0
click==8.3.0
colorlog
crewai[tools
cryptography==42.0.2
dataclasses-json>=0.6.0
datasets>=2.14.0
fastapi>=0.115.14
flake8 # For linting
flake8>=6.0.0
flask-cors==6.0.1
Flask==3.1.1
gitpython>=3.1.43
google-adk[a2a]==1.15.1
google-genai>=1.21.1
google-generativeai==0.3.0
httpx # For making async HTTP requests (e.g., in tests or if your app calls other services)
httpx-sse==0.4.3
httpx==0.28.1
huggingface-hub>=0.16.0
ipywidgets>=8.0.0
isort==5.13.2
jsonschema>=4.20.0
jupyter>=1.0.0
langchain-core>=0.1.0
langchain-mcp-adapters>=0.1.7
langchain-openai>=0.3.25
langchain>=0.3.26
langgraph>=0.4.9
litellm>=1.73.0
markdownify>=0.13.1
matplotlib>=3.5.0
mcp==1.21.2
mem0ai>=0.1.117
microsoft-agents-copilotstudio-client>=0.3.1
mypy>=1.7.0
nest-asyncio>=1.6.0
networkx==3.1
numpy==1.26.4
openai-chatkit>=1.1.0
openai>=1.88.0
opencv-python>=4.8.0
openpyxl>=3.1.0
opentelemetry-api>=1.24.0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-semantic-conventions-ai>=0.4.13
pandas==2.1.1
pennylane>=0.32.0
pgvector # Python client for pgvector PostgreSQL extension (for storing/querying vectors)
Pillow>=10.0.0
playwright==1.52.0
plotly>=5.0.0
protego>=0.3.1
psutil==5.9.6
psycopg2-binary # For connecting to PostgreSQL
psycopg2-binary>=2.9.0
pycryptodome==3.19.0
pydantic-settings # For loading settings from .env files or environment (replaces pydantic.BaseSettings in Pydantic v2)
pydantic-settings==2.12.0
pydantic==2.12.4
PyJWT[crypto]==2.10.1
PyPDF2>=3.0.0
pyright>=1.1.402
pytesseract>=0.3.10
pytest # For running tests
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest==9.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml"
    # requests? (if we add fetchers later)
PyYAML==6.0.3
qiskit-aer>=0.12.0
qiskit-ibm-provider>=0.7.0
qiskit>=0.45.0
rapidfuzz # For fast fuzzy string matching (alternative: fuzzywuzzy)
readabilipy>=0.2.0
redis>=6.4.0
redisvl>=0.8.2
requests==2.31.0
rich==13.7.0
schedule
scipy>=1.9.0
seaborn>=0.12.0
sentence-transformers # For generating text embeddings
sentence-transformers>=2.2.0
simhash # For SimHash fingerprinting
sqlalchemy==2.0.44
streamlit>=1.31.0
structlog==23.2.0
termcolor==3.1.0
torch>=2.0.0
tqdm>=4.65.0
transformers>=4.30.0
typing-extensions==4.9.0
tzdata>=2024.2
tzlocal>=5.3.1
uuid==1.30
uvicorn[standard]>=0.22.0 # Includes standard dependencies like watchfiles for --reload
watchdog>=4.0.0
websockets==12.0
youtube_search
